cleaned widgets and improved performance
