+++ MarvelWidgets/views/projectviews/ProjectListView.swift
@@ -13,6 +13,9 @@ struct ProjectListView: View {
     @StateObject var viewModel = ProjectListViewModel()
     @Binding var showLoader: Bool
     
+    @State var beforeDate: Date = Date()
+    @State var afterDate: Date = Date()
+    
     var body: some View {
         VStack{
             if viewModel.showFilters {
@@ -91,6 +94,26 @@ struct ProjectListView: View {
                                 }
                             }
                         }
+                    }
+                    
+                    HStack {
+                        Text("After date:")
+                        
+                        Spacer()
+                        
+                        DatePicker(selection: $beforeDate, displayedComponents: .date) {
+                                        Text("")
+                                    }
+                    }
+                    
+                    HStack {
+                        Text("Before date:")
+                        
+                        Spacer()
+                        
+                        DatePicker(selection: $afterDate, displayedComponents: .date) {
+                                        Text("")
+                                    }
                     }
