+++ SmallWidget/SmallWidget.swift
@@ -9,97 +9,6 @@ import WidgetKit
 import SwiftUI
 import Intents
 
-struct SmallWidgetProvider: IntentTimelineProvider {
-    let emptyProject = ProjectWrapper(
-        id: -1,
-        attributes: MCUProject(
-            title: "Next project",
-            releaseDate: Date.now.addingTimeInterval(60 * 60 * 24 * 2).ISO8601Format(),
-            postCreditScenes: nil,
-            duration: nil,
-            phase: .unkown,
-            saga: .infinitySaga,
-            overview: nil,
-            type: .special,
-            boxOffice: nil,
-            createdAt: nil,
-            updatedAt: nil,
-            disneyPlusUrl: nil,
-            directors: nil,
-            actors: nil,
-            relatedProjects: nil,
-            trailers: nil,
-            posters: nil,
-            seasons: nil
-        )
-    )
-    
-    func placeholder(in context: Context) -> UpcomingProjectEntry {
-        return UpcomingProjectEntry(
-            date: Date(),
-            configuration: WidgetTypeConfigurationIntent(),
-            upcomingProject: emptyProject,
-            nextProject: emptyProject,
-            image: Image("secret wars"),
-            nextImage: Image("secret wars")
-        )
-    }
-
-    func getSnapshot(for configuration: WidgetTypeConfigurationIntent, in context: Context, completion: @escaping (UpcomingProjectEntry) -> ()) {
-        let entry = UpcomingProjectEntry(
-            date: Date(),
-            configuration: WidgetTypeConfigurationIntent(),
-            upcomingProject: emptyProject,
-            nextProject: emptyProject,
-            image: Image("secret wars"),
-            nextImage: Image("secret wars")
-        )
-        
-        completion(entry)
-    }
-
-    func getTimeline(for configuration: WidgetTypeConfigurationIntent, in context: Context, completion: @escaping (Timeline<UpcomingProjectEntry>) -> ()) {
-        Task {
-            var entries: [UpcomingProjectEntry] = []
-            var upcomingProjects : [ProjectWrapper] = []
-            
-            let widgetType: WidgetType = WidgetType.getFromIndex(configuration.WidgetType.rawValue)
-            
-            switch widgetType {
-            case .movies:
-                upcomingProjects.append(contentsOf: await ProjectService.getByType(.movies, populate: .populateNormal))
-            case .series:
-                upcomingProjects.append(contentsOf: await ProjectService.getByType(.series, populate: .populateNormal))
-            case .special:
-                upcomingProjects.append(contentsOf: await ProjectService.getByType(.special, populate: .populateNormal))
-            case .all:
-                upcomingProjects.append(contentsOf: await ProjectService.getAll(populate: .populateNormal))
-            case .saved:
-                upcomingProjects.append(contentsOf: SaveService.getProjectsFromUserDefaults())
-            }
-            
-            switch configuration.RandomOrNext.rawValue {
-            case 2:
-                entries.append(WidgetProjectService.randomProject(from: upcomingProjects, with: configuration))
-            default:
-                entries.append(WidgetProjectService.upcomingProject(from: upcomingProjects, with: configuration))
-            }
-            
-            let timeline = Timeline(entries: entries, policy: .atEnd)
-            completion(timeline)
-        }
-    }
-}
-
-struct UpcomingProjectEntry: TimelineEntry {
-    let date: Date
-    let configuration: WidgetTypeConfigurationIntent
-    let upcomingProject: ProjectWrapper?
-    let nextProject: ProjectWrapper?
-    let image: Image
-    let nextImage: Image?
-}
-
 struct SmallWidgetUpcoming : View {
     @Environment(\.widgetFamily) var family
     var entry: SmallWidgetProvider.Entry
